+++ zoomimage-core-coil/src/main/java/com/github/panpf/zoomimage/coil/internal/CoilOtherUtils.kt
@@ -1,0 +1,4 @@
+package com.github.panpf.zoomimage.coil.internal
+
+
+internal fun Any.toHexString(): String = Integer.toHexString(this.hashCode())