+++ zoomimage-core-picasso/src/main/java/com/github/panpf/zoomimage/picasso/internal/PicassoOtherUtils.kt
@@ -1,0 +1,4 @@
+package com.github.panpf.zoomimage.picasso.internal
+
+
+internal fun Any.toHexString(): String = Integer.toHexString(this.hashCode())