+++ zoomimage-core-sketch/src/main/java/com/github/panpf/zoomimage/sketch/SketchTileBitmap.kt
@@ -18,6 +18,7 @@ package com.github.panpf.zoomimage.sketch
 
 import android.graphics.Bitmap
 import com.github.panpf.sketch.cache.MemoryCache
+import com.github.panpf.zoomimage.sketch.internal.toHexString
 import com.github.panpf.zoomimage.subsampling.AndroidCacheTileBitmap
 
 class SketchTileBitmap constructor(
@@ -38,11 +39,18 @@ class SketchTileBitmap constructor(
     override val isRecycled: Boolean
         get() = bitmap?.isRecycled ?: true
 
+    private val toString =
+        "SketchTileBitmap(size=${width}x${height},config=${bitmap!!.config},@${bitmap!!.toHexString()})"
+
     override fun recycle() {
         bitmap?.recycle()
     }
 
     override fun setIsDisplayed(displayed: Boolean) {
         cacheValue.countBitmap.setIsDisplayed(displayed, caller)
+    }
+
+    override fun toString(): String {
+        return toString
     }
 }